Update README.md
"The Walking Dead" is a popular post-apocalyptic horror franchise that explores a world ravaged by a zombie apocalypse. The story centers around a group of survivors navigating the dangers of a world overrun by flesh-eating undead, known as "walkers."
"American Psycho" has been both celebrated and condemned for its unflinching portrayal of violence and its critique of American society. The novel remains a powerful and disturbing commentary on the dark side of human nature and the seductive allure of materialism.
The American Civil War, a tumultuous period in American history, was a conflict primarily fought between the United States of America and the Confederate States of America. It spanned from 1861 to 1865, leaving a lasting impact on the nation's social, political, and economic fabric.