Update README.md
"The Walking Dead" is a popular post-apocalyptic horror franchise that explores a world ravaged by a zombie apocalypse. The story centers around a group of survivors navigating the dangers of a world overrun by flesh-eating undead, known as "walkers."
"American Psycho" is a dark and disturbing novel by Bret Easton Ellis, later adapted into a controversial film. The story follows Patrick Bateman, a wealthy and successful investment banker in 1980s New York City, who leads a double life as a serial killer.
The American Civil War, a tumultuous period in American history, was a conflict primarily fought between the United States of America and the Confederate States of America. It spanned from 1861 to 1865, leaving a lasting impact on the nation's social, political, and economic fabric.